Use utils.types.ts files for util types
