+++ server/src/routes/TradesRouter/Polygon/utils.types.ts
@@ -1,0 +1,18 @@
+import type { DBPolygonDailyOpenCloseModel, PolygonDailyOpenClose } from './types';
+
+export type CreateNoDataDailyOpenCloseProps = {
+  date: string;
+};
+
+export type CreateDBModelFromPolygonDailyOpenCloseProps = {
+  polygonDailyOpenClose: PolygonDailyOpenClose;
+  date: string;
+};
+
+export type CreateMarketDailyOpenCloseEntryFromDBModelProps = {
+  dbPolygonDailyOpenCloseModel: DBPolygonDailyOpenCloseModel;
+};
+
+export type ParseDateStringWithNewYorkTzProps = {
+  date: string;
+};
