+++ server/src/routes/TradesRouter/utils.types.ts
@@ -1,0 +1,8 @@
+export type NullableDataToFlatArrayProps = {
+  data: any;
+};
+
+export type ConvertDateFromNewYorkTzToLocalProps = {
+  date: string;
+  format: string;
+};
